requirements.txt
astroid==2.2.5
autopep8==1.4.3
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
Django==2.1.7
django-datatables-view==1.17.0
django-environ==0.4.5
django-chartjs==1.5.0
google-api-core==1.9.0
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.9
idna==2.8
isort==4.3.16
lazy-object-proxy==1.3.1
mccabe==0.6.1
pep8==1.7.1
protobuf==3.7.1
psycopg2-binary==2.7.7
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.5.0
pylint==2.3.1
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
sweetify==1.0.3
typed-ast==1.3.1
urllib3==1.24.1
wrapt==1.11.1
